requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lyricsgenius==3.3.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.6
spotipy==2.25.1
SQLAlchemy==2.0.40
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
Wikipedia-API==0.8.1
